requirements.txt
django == 5.0.6
djangorestframework == 3.15.1
drf_spectacular == 0.27.2
django-filter == 24.2
django-simple-history == 3.7.0
pytest == 8.2.2
pytest-django == 4.8.0
psycopg2 == 2.9.9
gunicorn == 22.0.0
ruff == 0.4.9